finished towers of hanoi WB 14
